requirements.txt
numpy==1.22.1
pygame==2.1.2
matplotlib
seaborn
pyyaml